drivers: adbms6830b: reorganize `service.rs` and `manager.rs` into `turnkey` module
